+++ tests/test_changelog.py
@@ -1,0 +1,18 @@
+#!/usr/bin/env python
+
+import unittest
+import mock
+
+from chag import changelog
+
+
+CHANGELOGA = """"""
+CHANGELOGB = """"""
+
+class TestChangelog(unittest.TestCase):
+    def test_ensures_file_exists(self):
+        pass
+
+
+if __name__ == '__main__':
+    unittest.main()