+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ CHANGELOG
 Next Release
 ------------
 
-Rewritten in Python using Click 3.0
+Rewritten in Python using Click.
 
 0.2.0 (2014-08-11)
 ------------------